# docs/swagger.yaml
basePath: /
definitions:
  github_com_luizhen1_go-clean-gin-crud_internal_domain.Product:
    description: Informações do produto
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API de exemplo de CRUD utilizando Clean Architecture e Gin.
  title: Go Clean Arch CRUD API
  version: "1.0"
paths:
  /products:
    get:
      description: Retorna uma lista de todos os produtos cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
            type: array
        "500":
          description: Erro ao buscar produtos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar produtos
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Cria um novo produto no banco de dados
      parameters:
      - description: Dados do Produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
        "400":
          description: Erro nos dados enviados
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno no servidor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar produto
      tags:
      - products
  /products/{id}:
    delete:
      description: Remove um produto do banco de dados
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao deletar
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar produto
      tags:
      - products
    get:
      description: Retorna um único produto pelo ID
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Produto não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar por ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um produto existente
      parameters:
      - description: ID do Produto
        in: path
        name: id
        required: true
        type: integer
      - description: Novos dados
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_luizhen1_go-clean-gin-crud_internal_domain.Product'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao atualizar
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar produto
      tags:
      - products
swagger: "2.0"